==> 02_pdr_analyses.Rmd <==
---
title: "PDR_Analyses"
author: "Jessica Kosie"
date: "3.27.20"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse); library(lme4); library(lmerTest); library(here); library(lsmeans)

source("helper/functions.R")

# make sure summary() uses Type III sums of squares
afex::set_sum_contrasts()

# turn off scientific notation
options(scipen = 999)


```

## Read in PDR data and remove fussy trials / blocks

```{r read in data}
pdr_all_trials <- read_csv("pdr_data/pdr_cleaned_trials_numbered.csv")

#remove fussy trials and blocks
remove <- read_csv("metadata/trials_to_remove.csv")

pdr_all_trials <- left_join(pdr_all_trials, remove)

#quick question - were more trials removed for infant- or adult-directed demos?
rem_compare <- pdr_all_trials %>% 
  filter(status == "remove") %>% 
  dplyr::select(vid, subid, remove_reason) %>% 
  unique()

rem_compare <- rem_compare %>% 
  separate(vid, c("actor", "partner", "object"))

table(rem_compare$partner, rem_compare$remove_reason)

table(rem_compare$partner)

#get rid of all trials that I decided to remove (because trials to keep only added "remove" when we want to remove trials and everything else is NA, keep only NAs in the `remove` column)
pdr <- pdr_all_trials %>% 
  filter(is.na(status))

#get rid of the extra data frame and extra columns    
rm(remove)
pdr$status <- NULL
pdr$remove_reason <- NULL

#remove ki009 and dr033 - they have a developmental delay
pdr <- pdr %>% 
  filter(subid != "ki009" & subid != "dr033")

#get rid of `trial_offset`
pdr <- pdr %>%
  filter(event != "trial_offset")



```
#Prepare for analyses.
```{r prep for analyses}
#get a boundary column and change boundary in the event column to `event_onset`
pdr <- pdr %>% 
  mutate(boundary = case_when(event == "boundary" ~ "boundary",
                              TRUE ~ "nonboundary"))

#doing this becuse boundaries are part of the event, and want to make it so that this is reflected in this column (e.g., separate event and boundary columns)
pdr <- pdr %>% 
  mutate(event = case_when(event == "boundary" ~ "event_onset",
                           TRUE ~ event))

#create a looking column (has a 1 when there is a raw_pdr value and 0 when raw_pdr is NA)
pdr <- pdr %>% 
  mutate(looking = case_when(!is.na(raw_pdr) ~ 1,
                           TRUE ~ 0))

#make actor and partner columns
pdr <- pdr %>% 
    separate(vid, into = c("actor","partner", "object"), sep = "_", remove = FALSE)
```
## Z-score PDR data. For each participant, get the mean and standard deviation across all of their blocks/trials and use those to caluclate pdr_z.
```{r z-score}
z_data <- pdr %>% 
  group_by(subid) %>% 
  dplyr::summarise(sub_mean = mean(raw_pdr, na.rm = TRUE),
                   sub_sd = sd(raw_pdr, na.rm = TRUE))

pdr <- pdr %>% 
  left_join(z_data)

pdr <- pdr %>% 
  mutate(pdr_z = (raw_pdr - sub_mean)/sub_sd)

rm(z_data)

```
#Filter z-scored data and get baseline values
```{r filter and baseline}
pdr <- pdr %>% 
  mutate(pdr_z_filtered = hanning_filter(pdr_z, 11))

#baseline correct
#calculate baseline PDR (last 1s of still before video begins playing)
baseline <- pdr %>% 
  filter(frame_num >= 150 & frame_num <= 180) %>% 
  group_by(subid, block, inBlock_trial_number) %>% 
  dplyr::summarise(baseline_pdr = mean(pdr_z_filtered))

pdr <- left_join(pdr, baseline)

#look at the distribution
hist(baseline$baseline_pdr)

#remove extraneous data file
rm(baseline)

#get only region of activity where the event (video) is happening
video_pdr <- filter(pdr, event == "event_onset")
```
#1. Is LOOKING greater to motionese or adult-directed action?
```{r looking to motionese vs ada}
vid_looking <- video_pdr %>% 
  group_by(subid, block, inBlock_trial_number, vid) %>% 
  count(looking)

#read in video lengths
vid_length <- read_csv("metadata/vid_lengths_per_matlab.csv")

#calculate video lengths without grey screen and still frame
vid_length <- vid_length %>% 
  mutate(vid_only_frames = vid_num_frames - 180)

#join with video looking file so that I can calculate prop looking to each video
vid_looking <- vid_looking %>% 
  filter(looking == 1) %>% 
  left_join(vid_length[,c(1,4)]) %>% 
  mutate(prop_looking = n / vid_only_frames)

vid_looking <- vid_looking %>% 
  separate(vid, into = c("actor", "partner", "object"), remove = FALSE)

#run analysis examining looking by interaction partner
look_mod <- lmer(prop_looking ~ partner + (1|subid) + (1|vid), data = vid_looking)
summary(look_mod)

#means and sds
vid_looking %>% 
  group_by(partner) %>% 
  summarise(mean_looking = mean(prop_looking),
            sd_looking = sd(prop_looking)) 

vid_looking$block <- as.factor(vid_looking$block) 

contrasts(vid_looking$block) <- contr.poly

#what if I add block?
look_mod <- lmer(prop_looking ~ partner * block + (1|subid), data = vid_looking)
summary(look_mod)
anova(look_mod)

#plot effect alone and then by block
plot.data <- vid_looking %>% 
  group_by(partner) %>% 
  summarise(mean = mean(prop_looking, na.rm=TRUE),
            sd = sd(prop_looking, na.rm=TRUE),
            n = n(),
            se=sd/sqrt(n),
            ci = qt(0.975,df=n-1)*se)

dodge <- position_dodge(width=0.9)

plot.data$partner <- factor(plot.data$partner, levels = c("Infant", "Adult"), labels = c("Motionese", "Adult-Directed"))

ggplot(plot.data, aes(x=partner, y=mean, fill=partner)) +
  geom_bar(stat="identity", position=position_dodge()) +
  geom_errorbar(aes(ymin=mean-ci, ymax=mean+ci),position=dodge, width=.1) +
  theme_linedraw() +
  theme(text = element_text(size=18), axis.text.x = element_text(size=18), axis.text.y = element_text(size=18), legend.position = "none") +
  labs(x = "Demonstration Type", y = "Average Proportion of Trial Spent Looking to Video")

#add blocks
plot.data <- vid_looking %>% 
  group_by(partner, block) %>% 
  summarise(mean = mean(prop_looking, na.rm=TRUE),
            sd = sd(prop_looking, na.rm=TRUE),
            n = n(),
            se=sd/sqrt(n),
            ci = qt(0.975,df=n-1)*se)

dodge <- position_dodge(width=0.9)

plot.data$partner <- factor(plot.data$partner, levels = c("Infant", "Adult"), labels = c("Motionese", "Adult-Directed"))

ggplot(plot.data, aes(x=as.factor(block), y=mean, fill=partner)) +
  geom_bar(stat="identity", position=position_dodge()) +
  geom_errorbar(aes(ymin=mean-ci, ymax=mean+ci),position=dodge, width=.1) +
  theme_linedraw() +
  theme(text = element_text(size=18), axis.text.x = element_text(size=18), axis.text.y = element_text(size=18)) +
  labs(x = "Block", y = "Average Proportion of Trial Spent Looking to Video", fill = "Demonstration Type")

#as line graph
ggplot(plot.data, aes(x = block, y = mean, group = partner, color = partner)) +
  geom_line(size = 1.5) +
  geom_point() +
  geom_errorbar(aes(ymin = mean-se, ymax = mean+se), width = .2, position = position_dodge(0.05), size = 1.5)+
  theme_linedraw() +
  theme(text = element_text(size=18), axis.text.x = element_text(size=18), axis.text.y = element_text(size=18)) +
  labs(x = "Block", y = "Average Proportion of Trial Spent Looking", fill = "Demonstration Type", color = "Demonstration Type")


# What if I z-score proportion looking?
vid_look_mean_sd <- vid_looking %>% 
  group_by(subid) %>% 
  summarise(mean_looking = mean(prop_looking),
            sd_looking = sd(prop_looking))

vid_looking <- left_join(vid_looking, vid_look_mean_sd)

vid_looking <- vid_looking %>% 
  mutate(prop_looking_z = (prop_looking - mean_looking)/sd_looking)

hist(vid_looking$prop_looking_z)

#summary
vid_looking %>% 
  group_by(partner) %>% 
  summarise(mean = mean(prop_looking_z, na.rm = TRUE),
            sd = sd(prop_looking_z, na.rm = TRUE))

#without block
look_mod <- lmer(prop_looking_z ~ partner + (1|subid), data = vid_looking)
summary(look_mod)
anova(look_mod)

#with block
look_mod <- lmer(prop_looking_z ~ partner * block + (1|subid), data = vid_looking)
summary(look_mod)
anova(look_mod)

#summary with block
vid_looking %>% 
  group_by(block, partner) %>% 
  summarise(mean = mean(prop_looking_z, na.rm = TRUE),
            sd = sd(prop_looking_z, na.rm = TRUE))

#binomial test for looking - 6/6 have greaer looking to infant-directed action
binom.test(6,6,p = .5,alternative="two.sided")

#plot for z-scored with blocks
plot.data <- vid_looking %>% 
  group_by(partner, block) %>% 
  summarise(mean = mean(prop_looking_z, na.rm=TRUE),
            sd = sd(prop_looking_z, na.rm=TRUE),
            n = n(),
            se=sd/sqrt(n),
            ci = qt(0.975,df=n-1)*se)

dodge <- position_dodge(width=0.9)

plot.data$partner <- factor(plot.data$partner, levels = c("Infant", "Adult"), labels = c("Motionese", "Adult-Directed"))

ggplot(plot.data, aes(x=as.factor(block), y=mean, fill=partner)) +
  geom_bar(stat="identity", position=position_dodge()) +
  geom_errorbar(aes(ymin=mean-ci, ymax=mean+ci),position=dodge, width=.1) +
  theme_linedraw() +
  theme(text = element_text(size=18), axis.text.x = element_text(size=18), axis.text.y = element_text(size=18)) +
  labs(x = "Block", y = "Average Proportion of Trial Spent Looking to Video", fill = "Demonstration Type")

#as line graph
ggplot(plot.data, aes(x = block, y = mean, group = partner, color = partner)) +
  geom_line(size = 1.5) +
  geom_point() +
  geom_errorbar(aes(ymin = mean-se, ymax = mean+se), width = .2, position = position_dodge(0.05), size = 1.5)+
  theme_linedraw() +
  theme(text = element_text(size=18), axis.text.x = element_text(size=18), axis.text.y = element_text(size=18)) +
  labs(x = "Block", y = "Average Proportion of Trial Spent Looking", fill = "Demonstration Type", color = "Demonstration Type")

```
#2. Is PDR greater to motionese or adult-directed action?
```{r PDR to motionese vs ada}
#run model predicting pdr_z_filtered from partner and baseline
model <- lmer(pdr_z_filtered ~ partner + baseline_pdr + (1|subid) + (1|vid), data = video_pdr)
summary(model)

video_pdr %>% 
  group_by(partner) %>% 
  summarise(mean_pdr = mean(pdr_z_filtered),
            sd_pdr = sd(pdr_z_filtered))

plot.data <- video_pdr %>% 
  group_by(frame_num, vid, object, partner) %>% 
  summarise(avg_pdr = mean(pdr_z_filtered, na.rm=TRUE),
            sd = sd(pdr_z_filtered, na.rm=TRUE),
            n = n(),
            se=sd/sqrt(n),
            ci = qt(0.975,df=n-1)*se) %>%   
  mutate(frame_num_from1 = frame_num - 180)

bounds <- read_csv("metadata/boundary_locations.csv")

bounds <- bounds %>% 
  separate(video, c("actor", "partner", "object")) %>% 
  mutate(frame_num_from1 = frame_num - 180)

plot.data$partner <- factor(plot.data$partner, levels = c("Infant", "Adult"), labels = c("Motionese", "Adult-Directed"))
 
bounds$partner <- factor(bounds$partner, levels = c("Infant", "Adult"), labels = c("Motionese", "Adult-Directed"))

plot.data %>% 
  ggplot(aes(x = frame_num_from1, y = avg_pdr, color = partner)) +
  geom_line(size = 1) +
  geom_ribbon(aes(ymin=avg_pdr-se, ymax=avg_pdr+se, fill = partner), alpha=0.4, show.legend = FALSE, colour = NA) +
  labs(y = "Average Z-Scored and Filtered Pupil Size", x = "Frame Number", color = "Demonstration Type") +
  facet_wrap(~object) +
  geom_vline(data = bounds, mapping = aes(xintercept = frame_num_from1, color = partner), size = 1) +
  theme_linedraw() +
  theme(text = element_text(size=18), axis.text.x = element_text(size=18), axis.text.y = element_text(size=18)) 

#what if you include block?
model_block <- lmer(pdr_z_filtered ~ partner * as.factor(block) + baseline_pdr + (1|subid) + (1|vid), data = video_pdr)
summary(model_block)
anova(model_block)

#overall
plot.data <- video_pdr %>% 
  group_by(frame_num, partner) %>% 
  dplyr::summarise(avg_pdr = mean(pdr_z_filtered, na.rm = TRUE)) %>% 
  mutate(frame_num_from1 = frame_num - 180)

plot.data %>% 
  ggplot(aes(x = frame_num_from1, y = avg_pdr, color = partner)) +
  geom_line(size = 1) +
  labs(y = "Average Z-Scored and Filtered PDR", x = "Frame Number", color = "Interaction Partner") +
  theme_linedraw() +
  theme(text = element_text(size=18), axis.text.x = element_text(size=18), axis.text.y = element_text(size=18)) 

#now with block
plot.data <- video_pdr %>% 
  group_by(frame_num, partner, block) %>% 
  dplyr::summarise(avg_pdr = mean(pdr_z_filtered, na.rm = TRUE)) %>% 
  mutate(frame_num_from1 = frame_num - 180)

plot.data %>% 
  ggplot(aes(x = frame_num_from1, y = avg_pdr, color = partner)) +
  geom_line(size = 1) +
  labs(y = "Average Z-Scored and Filtered PDR", x = "Frame Number", color = "Interaction Partner") +
  theme_linedraw() +
  facet_wrap(~block, nrow = 2) +
  theme(text = element_text(size=18), axis.text.x = element_text(size=18), axis.text.y = element_text(size=18)) 

mod1 <- lmer(pdr_z_filtered ~ partner + baseline_pdr + (1|subid) + (1|vid), data = filter(video_pdr, block == 1))
summary(mod1)

mod2 <- lmer(pdr_z_filtered ~ partner + baseline_pdr + (1|subid) + (1|vid), data = filter(video_pdr, block == 2))
summary(mod2)

mod3 <- lmer(pdr_z_filtered ~ partner + baseline_pdr + (1|subid) + (1|vid), data = filter(video_pdr, block == 3))
summary(mod3)

mod4 <- lmer(pdr_z_filtered ~ partner + baseline_pdr + (1|subid) + (1|vid), data = filter(video_pdr, block == 4))
summary(mod4)

mod5 <- lmer(pdr_z_filtered ~ partner + baseline_pdr + (1|subid) + (1|vid), data = filter(video_pdr, block == 5))
summary(mod5)

mod6 <- lmer(pdr_z_filtered ~ partner + baseline_pdr + (1|subid) + (1|vid), data = filter(video_pdr, block == 6))
summary(mod6)

video_pdr %>% 
  group_by(block, partner) %>% 
  summarise(mean = mean(pdr_z_filtered))

```
#2. Is infants' PDR greater at boundaries than pre- or post-boundary regions?
```{r PDR to boundary regions}

#import boundary regions file
bounds_1s <- read_csv("metadata/boundary_regions_1000ms.csv")

#join with PDR file to keep only boundary regions
bounds_1s <- left_join(bounds_1s, video_pdr)

#add object/partner
bounds_1s <- bounds_1s %>% 
  separate(vid, into = c("actor", "partner", "object"), sep = "_", remove = FALSE)

bounds_1s$region_1s <- factor(bounds_1s$region_1s, levels = c("pre_boundary", "boundary", "post_boundary"))

contrasts(bounds_1s$region_1s) <- contr.poly

#check that the contrasts look right
attributes(bounds_1s$region_1s)

model <- lmer(pdr_z_filtered ~ region_1s + baseline_pdr + (1|subid) + (1|vid), data = bounds_1s)
summary(model)
anova(model)

est.marginal.means <- lsmeans(model, specs="region_1s")
est.marginal.means
pairs(est.marginal.means, adjust = "bonferroni")

#means
bounds_1s %>% 
  group_by(region_1s) %>% 
  summarise(mean = mean(pdr_z_filtered, na.rm = TRUE),
            sd = sd(pdr_z_filtered, na.rm = TRUE))

#plot this effect 
plot.data <- bounds_1s %>% 
  group_by(reg_num, region_1s) %>% 
  dplyr::summarise(avg_pdr = mean(pdr_z_filtered, na.rm = TRUE))

plot.data %>% 
  ggplot(aes(x = reg_num, y = avg_pdr)) +
  geom_line(size = 1) +
  labs(title = "PDR to Boundaries", x = "Frames; Boundary at 0", y = "Average Z-Scored, Filtered PDR") +
  geom_vline(xintercept = 0, linetype = "dashed", size = 1) +
  geom_vline(xintercept = c(-30, 29, 59), color = "red", size = 1)

#plot with color shaded background
plot.data <- bounds_1s %>% 
  group_by(region_1s, reg_num) %>% 
  summarise(mean = mean(pdr_z_filtered, na.rm = TRUE))

rects <- data.frame(xstart = c(-30, 0, 29), xend = c(0, 29, 59), Region = c("pre-boundary", "boundary", "post-boundary"))

rects <- rects %>% 
  mutate(Region = factor(Region, levels = c("pre-boundary", "boundary", "post-boundary")))

#first, just rects without line
ggplot(plot.data, aes(x = reg_num, y = mean)) +
  geom_rect(data = rects, aes(xmin = xstart, xmax = xend, ymin = -Inf, ymax = Inf, fill = Region), alpha = 0.6, inherit.aes = FALSE) +
  ylim(-0.15, 0.10) +
  theme_linedraw(base_size = 24) +
  geom_vline(xintercept = 0, color = "forestgreen", linetype = "dashed", size = 2) +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(y = "Average Z-Scored and Filtered Pupil Size", x = "Frame Number (Boundary at 0)", linetype = "Interaction Partner", color = "Video")

#now add the line
ggplot(plot.data, aes(x = reg_num, y = mean)) +
  geom_rect(data = rects, aes(xmin = xstart, xmax = xend, ymin = -Inf, ymax = Inf, fill = Region), alpha = 0.6, inherit.aes = FALSE) +
  geom_line(size = 2) +
  ylim(-0.15, 0.10) +
  theme_linedraw(base_size = 24) +
    geom_vline(xintercept = 0, color = "forestgreen", linetype = "dashed", size = 2) +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(y = "Average Z-Scored and Filtered Pupil Size", x = "Frame Number (Boundary at 0)", linetype = "Interaction Partner", color = "Video")

#plot with shading around line
plot.data <- bounds_1s %>% 
  group_by(region_1s, reg_num) %>% 
  summarise(mean = mean(pdr_z_filtered, na.rm = TRUE),
            sd = sd(pdr_z_filtered, na.rm=TRUE),
            n = n(),
            se=sd/sqrt(n),
            ci = qt(0.975,df=n-1)*se)

rects <- data.frame(xstart = c(-30, 0, 29), xend = c(0, 29, 59), Region = c("pre-boundary", "boundary", "post-boundary"))

rects <- rects %>% 
  mutate(Region = factor(Region, levels = c("pre-boundary", "boundary", "post-boundary")))

ggplot(plot.data, aes(x = reg_num, y = mean)) +
  geom_rect(data = rects, aes(xmin = xstart, xmax = xend, ymin = -Inf, ymax = Inf, fill = Region), alpha = 0.6, inherit.aes = FALSE) +
  geom_line(size = 2) +
  geom_ribbon(aes(ymin=mean-se, ymax=mean+se), alpha=0.4) +
  ylim(-0.15, 0.10) +
  theme_linedraw(base_size = 24) +
  geom_vline(xintercept = 0, color = "forestgreen", linetype = "dashed", size = 2) +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(y = "Average Z-Scored and Filtered Pupil Size", x = "Frame Number (Boundary at 0)", linetype = "Interaction Partner", color = "Video")


```
#3. Does motionese enhance infant's attention to event boundaries?
```{r boundary x motionese PDR}
model <- lmer(pdr_z_filtered ~ region_1s * partner + baseline_pdr + (1|subid) + (1|vid), data = bounds_1s)
summary(model)
anova(model)

model_adult <- lmer(pdr_z_filtered ~ region_1s + baseline_pdr + (1|subid) + (1|vid), data = filter(bounds_1s, partner == "Adult"))
summary(model_adult)
anova(model_adult)

est.marginal.means <- lsmeans(model_adult, specs="region_1s")
est.marginal.means
pairs(est.marginal.means, adjust = "bonferroni")

model_infant <- lmer(pdr_z_filtered ~ region_1s + baseline_pdr + (1|subid) + (1|vid), data = filter(bounds_1s, partner == "Infant"))
summary(model_infant)
anova(model_infant)

est.marginal.means <- lsmeans(model_infant, specs="region_1s")
est.marginal.means
pairs(est.marginal.means, adjust = "bonferroni")

bounds_1s %>% 
  filter(partner == "Infant") %>% 
  group_by(region_1s) %>% 
  summarise(mean_infant = mean(pdr_z_filtered),
            sd_infant = sd(pdr_z_filtered))

plot.data <- bounds_1s %>% 
  group_by(reg_num, region_1s, partner) %>% 
  dplyr::summarise(avg_pdr = mean(pdr_z_filtered, na.rm = TRUE))

plot.data %>% 
  ggplot(aes(x = reg_num, y = avg_pdr, color = partner)) +
  geom_line(size = 1) +
  labs(title = "PDR to Boundaries for Adult- vs Infant-Directed Action", x = "Frames; Boundary at 0", y = "Average Z-Scored, Filtered, Baseline Corrected PDR") +
  geom_vline(xintercept = 0, size = 1) +
  geom_vline(xintercept = c(-30, 29, 59), linetype = "dashed", size = 1)

#plot the same way as I did luminance
plot.data <- bounds_1s %>% 
  group_by(region_1s, partner, reg_num) %>% 
  summarise(mean = mean(pdr_z_filtered, na.rm = TRUE),
            sd = sd(pdr_z_filtered, na.rm=TRUE),
            n = n(),
            se=sd/sqrt(n),
            ci = qt(0.975,df=n-1)*se)

rects <- data.frame(xstart = c(-30, 0, 29), xend = c(0, 29, 59), Region = c("pre-boundary", "boundary", "post-boundary"))

rects <- rects %>% 
  mutate(Region = factor(Region, levels = c("pre-boundary", "boundary", "post-boundary")))

plot.data$partner <- factor(plot.data$partner, levels = c("Infant", "Adult"), labels = c("Motionese", "Adult-Directed"))

rects <- rects %>% 
  mutate(color = case_when(Region == "pre-boundary" ~ "#00BFC4",
                           Region == "boundary" ~ "#F8766D",
                           Region == "post-boundary" ~ "#C77CFF", 
                           TRUE ~ NA_character_))

ggplot(plot.data, aes(x = reg_num, y = mean, linetype = partner)) +
  geom_rect(data = rects, aes(xmin = xstart, xmax = xend, ymin = -Inf, ymax = Inf), fill = rects$color, alpha = 0.5, inherit.aes = FALSE) +
  geom_vline(xintercept = 0, color = "indianred2", linetype = "dashed", size = 2) +
  geom_line(size = 2) +
  geom_ribbon(aes(ymin=mean-se, ymax=mean+se), alpha=0.2) +
  ylim(-0.2, 0.15) +
  theme_linedraw(base_size = 18) +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(y = "Average Z-Scored and Filtered Pupil Size", x = "Frame Number (Boundary at 0)", linetype = "Demonstration Type") +
  guides(linetype = guide_legend(override.aes = list(size = 1))) +
  annotate("text", x = -15, y= 0.15, label = "Pre-Boundary Region", size = 5) +
  annotate("text", x = 15, y= 0.15, label = "Boundary Region", size = 5) +
  annotate("text", x = 45, y= 0.15, label = "Post-Boundary Region", size = 5) 

#with original colors
ggplot(plot.data, aes(x = reg_num, y = mean, linetype = partner)) +
  geom_rect(data = rects, aes(xmin = xstart, xmax = xend, ymin = -Inf, ymax = Inf, fill = Region), alpha = 0.6, inherit.aes = FALSE) +
  geom_line(size = 2) +
  geom_ribbon(aes(ymin=mean-se, ymax=mean+se), alpha=0.4) +
  ylim(-0.2, 0.15) +
  theme_linedraw(base_size = 24) +
  geom_vline(xintercept = 0, color = "forestgreen", linetype = "dashed", size = 2) +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(y = "Average Z-Scored and Filtered Pupil Size", x = "Frame Number (Boundary at 0)", linetype = "Demonstration Type") 
```
#3a. Does infants' response to boundaries in IDA differ across blocks?
```{r boundary x motionese PDR}
# add 3-way ineraction with block
model <- lmer(pdr_z_filtered ~ region_1s * partner * block + baseline_pdr + (1|subid) + (1|vid), data = bounds_1s)
summary(model)
anova(model)

bounds_1s$block <- as.factor(bounds_1s$block)
contrasts(bounds_1s$block) <- contr.poly

model_infant <- lmer(pdr_z_filtered ~ region_1s * as.factor(block) + baseline_pdr + (1|subid) + (1|vid), data = filter(bounds_1s, partner == "Infant"))
summary(model_infant)
anova(model_infant)

bounds_1s %>% 
  filter(partner == "Infant") %>% 
  group_by(region_1s, block) %>% 
  summarise(mean_infant = mean(pdr_z_filtered),
            sd_infant = sd(pdr_z_filtered))

#plot the same way as I did luminance
plot.data <- bounds_1s %>% 
  filter(partner == "Infant") %>% 
  group_by(region_1s, reg_num, block) %>% 
  summarise(mean = mean(pdr_z_filtered, na.rm = TRUE),
            sd = sd(pdr_z_filtered, na.rm=TRUE),
            n = n(),
            se=sd/sqrt(n),
            ci = qt(0.975,df=n-1)*se)

rects <- data.frame(xstart = c(-30, 0, 29), xend = c(0, 29, 59), Region = c("pre-boundary", "boundary", "post-boundary"))

rects <- rects %>% 
  mutate(Region = factor(Region, levels = c("pre-boundary", "boundary", "post-boundary")))

plot.data$block <- as.factor(plot.data$block)

ggplot(plot.data, aes(x = reg_num, y = mean, color = block)) +
  geom_rect(data = rects, aes(xmin = xstart, xmax = xend, ymin = -Inf, ymax = Inf, fill = Region), alpha = 0.2, inherit.aes = FALSE) +
  geom_line(size = 2) +
  geom_ribbon(aes(ymin=mean-se, ymax=mean+se, fill = block), alpha=0.4, show.legend = FALSE) +
  theme_linedraw(base_size = 16) +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(y = "Average Z-Scored and Filtered Pupil Size", x = "Frame Number (Boundary at 0)", linetype = "Interaction Partner", color = "Block")

plot.data <- plot.data %>% 
  mutate(block = factor(block, levels = c(1,2,3,4,5,6), labels = c("Block 1", "Block 2", "Block 3", "Block 4", "Block 5", "Block 6")))

# same plot but faceted by block
ggplot(plot.data, aes(x = reg_num, y = mean)) +
  geom_rect(data = rects, aes(xmin = xstart, xmax = xend, ymin = -Inf, ymax = Inf, fill = Region), alpha = 0.2, inherit.aes = FALSE) +
  geom_line(size = 2) +
  geom_ribbon(aes(ymin=mean-se, ymax=mean+se), alpha=0.4, show.legend = FALSE) +
  theme_linedraw(base_size = 16) +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(y = "Average Z-Scored and Filtered Pupil Size", x = "Frame Number (Boundary at 0)", linetype = "Interaction Partner", color = "Block") +
  facet_wrap(~block)


#separately by block
bl1 <- lmer(pdr_z_filtered ~ region_1s + baseline_pdr + (1|vid), data = filter(bounds_1s, partner == "Infant" & block == 1))
summary(bl1)
anova(bl1)

bl2 <- lmer(pdr_z_filtered ~ region_1s + baseline_pdr + (1|subid) + (1|vid), data = filter(bounds_1s, partner == "Infant" & block == 2))
summary(bl2)
anova(bl2)

bl3 <- lmer(pdr_z_filtered ~ region_1s + baseline_pdr + (1|subid) + (1|vid), data = filter(bounds_1s, partner == "Infant" & block == 3))
summary(bl3)

bl4 <- lmer(pdr_z_filtered ~ region_1s + baseline_pdr + (1|subid) + (1|vid), data = filter(bounds_1s, partner == "Infant" & block == 4))
summary(bl4)

est.marginal.means <- lsmeans(bl4, specs="region_1s")
est.marginal.means
pairs(est.marginal.means, adjust = "bonferroni")

bl5 <- lmer(pdr_z_filtered ~ region_1s + baseline_pdr + (1|subid), data = filter(bounds_1s, partner == "Infant" & block == 5))
summary(bl5)

est.marginal.means <- lsmeans(bl5, specs="region_1s")
est.marginal.means
pairs(est.marginal.means, adjust = "bonferroni")

bl6 <- lmer(pdr_z_filtered ~ region_1s + baseline_pdr + (1|subid) + (1|vid), data = filter(bounds_1s, partner == "Infant" & block == 6))
summary(bl6)

# will want to control for multiple comparisons here
# p.vals <- c(.183, .598, .35, .0356, .00562, .02804, .000536, .000434, .11286, .00115, .1078, .0614)
# 
# p.adjust(p.vals, method = "bonferroni")
```
#out of curiosity - plot by object (maybe this goes in the questions part of the presentation)
```{r}
#plot the same way as I did luminance (infant-directed only)
plot.data <- bounds_1s %>% 
  filter(partner == "Infant") %>% 
  group_by(region_1s, reg_num, vid) %>% 
  summarise(mean = mean(pdr_z_filtered))

rects <- data.frame(xstart = c(-30, 0, 29), xend = c(0, 29, 59), Region = c("pre-boundary", "boundary", "post-boundary"))

rects <- rects %>% 
  mutate(Region = factor(Region, levels = c("pre-boundary", "boundary", "post-boundary")))

ggplot(plot.data, aes(x = reg_num, y = mean, color = vid)) +
  geom_rect(data = rects, aes(xmin = xstart, xmax = xend, ymin = -Inf, ymax = Inf, fill = Region), alpha = 0.2, inherit.aes = FALSE) +
  geom_line(size = 2) +
  theme_linedraw(base_size = 16) +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(y = "Average Z-Scored and Filtered Pupil Size", x = "Frame Number (Boundary at 0)", linetype = "Interaction Partner", color = "Video")

#plot the same way as I did luminance - same for adult-directed
plot.data <- bounds_1s %>% 
  filter(partner == "Adult") %>% 
  group_by(region_1s, reg_num, vid) %>% 
  summarise(mean = mean(pdr_z_filtered))

rects <- data.frame(xstart = c(-30, 0, 29), xend = c(0, 29, 59), Region = c("pre-boundary", "boundary", "post-boundary"))

rects <- rects %>% 
  mutate(Region = factor(Region, levels = c("pre-boundary", "boundary", "post-boundary")))

ggplot(plot.data, aes(x = reg_num, y = mean, color = vid)) +
  geom_rect(data = rects, aes(xmin = xstart, xmax = xend, ymin = -Inf, ymax = Inf, fill = Region), alpha = 0.2, inherit.aes = FALSE) +
  geom_line(size = 2) +
  theme_linedraw(base_size = 16) +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(y = "Average Z-Scored and Filtered Pupil Size", x = "Frame Number (Boundary at 0)", linetype = "Interaction Partner", color = "Video")

```
#separate boundary plot for each subject
```{r}
#plot same as above but facet for subid
plot.data <- bounds_1s %>% 
  group_by(region_1s, partner, reg_num, subid) %>% 
  summarise(mean = mean(pdr_z_filtered)) %>% 
  mutate(subid = parse_number(subid))

rects <- data.frame(xstart = c(-30, 0, 29), xend = c(0, 29, 59), Region = c("pre-boundary", "boundary", "post-boundary"))

rects <- rects %>% 
  mutate(Region = factor(Region, levels = c("pre-boundary", "boundary", "post-boundary")))

plot.data$partner <- factor(plot.data$partner, levels = c("Infant", "Adult"), labels = c("Motionese", "Adult-Directed"))

ggplot(plot.data, aes(x = reg_num, y = mean, linetype = partner)) +
  geom_rect(data = rects, aes(xmin = xstart, xmax = xend, ymin = -Inf, ymax = Inf, fill = Region), alpha = 0.2, inherit.aes = FALSE) +
  geom_line() +
  theme_linedraw(base_size = 16) +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(y = "Average Z-Scored and Filtered Pupil Size", x = "Frame Number (Boundary at 0)", linetype = "Demonstration Type") +
  facet_wrap(~subid, nrow = 4)

#same with free scales
ggplot(plot.data, aes(x = reg_num, y = mean, linetype = partner)) +
  geom_rect(data = rects, aes(xmin = xstart, xmax = xend, ymin = -Inf, ymax = Inf, fill = Region), alpha = 0.2, inherit.aes = FALSE) +
  geom_line() +
  theme_linedraw(base_size = 14) +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(y = "Average Z-Scored and Filtered Pupil Size", x = "Frame Number (Boundary at 0)", linetype = "Demonstration Type") +
  facet_wrap(~subid, nrow = 4, scales = "free")


```
#4. Does luminance account for this observed effect?
```{r luminance}
#read in luminance file (new luminance values with corrected equation)
luminance <- read_csv('metadata/luminance.csv')

#remove subid from vid in luminance file
luminance <- luminance %>% 
  separate(vid, c("object", "partner")) 

#select only the columns we want to join with the data file
luminance <- luminance[,c("partner", "object", "frame_num", "lum_z")]

#join with boundary file
bounds_1s <- left_join(bounds_1s, luminance)

model <- lmer(pdr_z_filtered ~ region_1s * partner + baseline_pdr + lum_z + (1|subid) + (1|vid), data = bounds_1s)
summary(model)
anova(model)

model_adult <- lmer(pdr_z_filtered ~ region_1s + baseline_pdr + lum_z + (1|subid) + (1|vid), data = filter(bounds_1s, partner == "Adult"))
summary(model_adult)
anova(model_adult)

model_infant <- lmer(pdr_z_filtered ~ region_1s + baseline_pdr + lum_z + (1|subid) + (1|vid), data = filter(bounds_1s, partner == "Infant"))
summary(model_infant)
anova(model_infant)

#plot with luminance on same plot as PDR
#plot the same way as I did luminance
plot.data <- bounds_1s %>% 
  group_by(region_1s, partner, reg_num) %>% 
  summarise(mean_pdr = mean(pdr_z_filtered, na.rm=TRUE),
            sd_pdr = sd(pdr_z_filtered, na.rm=TRUE),
            n_pdr = n(),
            se_pdr=sd_pdr/sqrt(n_pdr),
            ci_pdr = qt(0.975,df=n_pdr-1)*se_pdr,
            mean_lum = mean(lum_z, na.rm=TRUE),
            sd_lum = sd(lum_z, na.rm=TRUE),
            n_lum = n(),
            se_lum=sd_lum/sqrt(n_lum),
            ci_lum = qt(0.975,df=n_lum-1)*se_lum) 

rects <- data.frame(xstart = c(-30, 0, 29), xend = c(0, 29, 59), Region = c("pre-boundary", "boundary", "post-boundary"))

rects <- rects %>% 
  mutate(Region = factor(Region, levels = c("pre-boundary", "boundary", "post-boundary")))

plot.data$partner <- factor(plot.data$partner, levels = c("Infant", "Adult"), labels = c("Motionese", "Adult-Directed"))

ggplot(plot.data) +
  geom_rect(data = rects, aes(xmin = xstart, xmax = xend, ymin = -Inf, ymax = Inf, fill = Region), alpha = 0.2, inherit.aes = FALSE) +
  geom_line(aes(x = reg_num, y = mean_pdr, linetype = partner), size = 2) +
  geom_ribbon(aes(x = reg_num, ymin=mean_pdr-se_pdr, ymax=mean_pdr+se_pdr, fill = partner), alpha=0.4, show.legend = FALSE, colour = NA) +
  geom_line(aes(x = reg_num, y = mean_lum, linetype = partner), color = "red", size = 2) +
  theme_linedraw(base_size = 16) +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(y = "Pupil Size (Black) & Luminance (Red)", x = "Frame Number (Boundary at 0)", linetype = "Demonstration Type") +
  guides(linetype = guide_legend(override.aes = list(size = 1)))

# Plot with different y axes

rects <- data.frame(xstart = c(-30, 0, 29), xend = c(0, 29, 59), Region = c("pre-boundary", "boundary", "post-boundary"))

rects <- rects %>% 
  mutate(Region = factor(Region, levels = c("pre-boundary", "boundary", "post-boundary")))

par(mar=c(5, 4, 4, 4) + 0.1)
plot(mean_pdr ~ reg_num, ann = FALSE, type = "l", data = filter(plot.data, partner == "Adult-Directed"), ylim = c(-0.15, 0.10), lty = 2, lwd = 3)
par(new = TRUE)
plot(mean_pdr ~ reg_num, ann = FALSE, type = "l", data = filter(plot.data, partner == "Motionese"), ylim = c(-0.15, 0.10), lwd = 3)
par(new = TRUE)
plot(mean_lum ~ reg_num, yaxt = "none", ann = FALSE, type = "l", data = filter(plot.data, partner == "Adult-Directed"), ylim = c(-0.7, 0.6), lty = 2, lwd = 3, col = "red")
par(new = TRUE)
plot(mean_lum ~ reg_num, yaxt = "none", ann = FALSE, type = "l", data = filter(plot.data, partner == "Motionese"), ylim = c(-0.7, 0.6), lwd = 3, col = "red", yaxt="none")
par(new = TRUE)
axis(4, seq(-0.7, 0.6, 0.3))
mtext(side=4, line=3, "Average Z-Scored Luminance")
mtext(side=2, line=3, "Average Z-Scored Pixel Size")
mtext(side=1, line=3, "Frame Number (Boundary at 0)")


```
#luminance across items
```{r luminance across items}
#Dare asked about luminance across the different items -- is it more problematic for some? I haven't done that analysis, but we can look quickly.
model <- lmer(pdr_z_filtered ~ region_1s * partner * object + baseline_pdr + lum_z + (1|subid) + (1|vid), data = bounds_1s)
summary(model)
anova(model)

#Do above analyses separately for each object
slinky <- lmer(pdr_z_filtered ~ region_1s * partner + baseline_pdr + lum_z + (1|subid) + (1|vid), data = filter(bounds_1s, object == "Slinky"))
anova(slinky)

ball <- lmer(pdr_z_filtered ~ region_1s * partner + baseline_pdr + lum_z + (1|subid) + (1|vid), data = filter(bounds_1s, object == "StickyBall"))
anova(ball)

glasses <- lmer(pdr_z_filtered ~ region_1s * partner + baseline_pdr + lum_z + (1|subid) + (1|vid), data = filter(bounds_1s, object == "TwistyGlasses"))
anova(glasses)

roller <- lmer(pdr_z_filtered ~ region_1s * partner + baseline_pdr + lum_z + (1|subid) + (1|vid), data = filter(bounds_1s, object == "MassageRoller"))
anova(roller)

tube <- lmer(pdr_z_filtered ~ region_1s * partner + baseline_pdr + lum_z + (1|subid) + (1|vid), data = filter(bounds_1s, object == "GreenTube"))
anova(tube)

stacker <- lmer(pdr_z_filtered ~ region_1s * partner + baseline_pdr + lum_z + (1|subid) + (1|vid), data = filter(bounds_1s, object == "OballStacker"))
anova(stacker)

#plot by object
#plot the same way as I did luminance
plot.data <- bounds_1s %>% 
  group_by(region_1s, reg_num, partner, object) %>% 
  summarise(mean = mean(pdr_z_filtered))

rects <- data.frame(xstart = c(-30, 0, 29), xend = c(0, 29, 59), Region = c("pre-boundary", "boundary", "post-boundary"))

rects <- rects %>% 
  mutate(Region = factor(Region, levels = c("pre-boundary", "boundary", "post-boundary")))

ggplot(plot.data, aes(x = reg_num, y = mean, linetype = partner)) +
  geom_rect(data = rects, aes(xmin = xstart, xmax = xend, ymin = -Inf, ymax = Inf, fill = Region), alpha = 0.2, inherit.aes = FALSE) +
  geom_line(size = 2) +
  theme_linedraw(base_size = 16) +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(y = "Average Z-Scored and Filtered Pupil Size", x = "Frame Number (Boundary at 0)", linetype = "Interaction Partner", color = "Video") +
  facet_wrap(~object)

```
#5. Same analysis controlling for infant age (plus looking analysis controlling for infant age)
```{r controlling for age}
#read in age file
age <- read_csv(here('chapter_3', 'age_days.csv'))

vid_looking <- left_join(vid_looking, age)

#run analysis examining looking by interaction partner
look_mod <- lmer(prop_looking ~ partner * block + age_days + (partner|subid) + (1|vid), data = vid_looking)
summary(look_mod)
anova(look_mod)

#join with boundary file
bounds_1s <- left_join(bounds_1s, age)

model <- lmer(pdr_z_filtered ~ region_1s * partner + age_days + baseline_pdr + lum_z + (1|subid) + (1|vid), data = bounds_1s)
summary(model)
anova(model)

bounds_1s <- bounds_1s %>% 
  mutate(age_group = case_when(age_days <= 287 ~ "youngest",
                               age_days > 287 & age_days < 322 ~ "middlest",
                               age_days >= 322 ~ "oldest", 
                               TRUE ~ NA_character_))

plot.data <- bounds_1s %>% 
  group_by(region_1s, partner, reg_num, age_group) %>% 
  summarise(mean_pdr = mean(pdr_z_filtered),
            mean_lum = mean(lum_z))

rects <- data.frame(xstart = c(-30, 0, 29), xend = c(0, 29, 59), Region = c("pre-boundary", "boundary", "post-boundary"))

rects <- rects %>% 
  mutate(Region = factor(Region, levels = c("pre-boundary", "boundary", "post-boundary")))

plot.data <- plot.data %>% 
  mutate(age_group = factor(age_group, levels = c("youngest", "middlest", "oldest")))

ggplot(plot.data) +
  geom_rect(data = rects, aes(xmin = xstart, xmax = xend, ymin = -Inf, ymax = Inf, fill = Region), alpha = 0.2, inherit.aes = FALSE) +
  geom_line(aes(x = reg_num, y = mean_pdr, linetype = partner), size = 2) +
  theme_linedraw(base_size = 16) +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(y = "Pupil Size (Black) & Luminance (Red)", x = "Frame Number (Boundary at 0)", linetype = "Interaction Partner") +
  guides(linetype = guide_legend(override.aes = list(size = 1))) +
  facet_wrap(~age_group)

#YOUNGEST
model <- lmer(pdr_z_filtered ~ region_1s * partner + baseline_pdr + lum_z + (1|subid) + (1|vid), data = filter(bounds_1s, age_group == "youngest"))
summary(model)
anova(model)

model <- lmer(pdr_z_filtered ~ region_1s + baseline_pdr + lum_z + (1|subid) + (1|vid), data = filter(bounds_1s, age_group == "youngest" & partner == "Infant"))
summary(model)
anova(model)

model <- lmer(pdr_z_filtered ~ region_1s + baseline_pdr + lum_z + (1|subid) + (1|vid), data = filter(bounds_1s, age_group == "youngest" & partner == "Adult"))
summary(model)
anova(model)

#MIDDLEST
model <- lmer(pdr_z_filtered ~ region_1s * partner + baseline_pdr + lum_z + (1|subid) + (1|vid), data = filter(bounds_1s, age_group == "middlest"))
summary(model)
anova(model)

model <- lmer(pdr_z_filtered ~ region_1s + baseline_pdr + lum_z + (1|subid) + (1|vid), data = filter(bounds_1s, age_group == "middlest" & partner == "Infant"))
summary(model)
anova(model)

model <- lmer(pdr_z_filtered ~ region_1s + baseline_pdr + lum_z + (1|subid) + (1|vid), data = filter(bounds_1s, age_group == "middlest" & partner == "Adult"))
summary(model)
anova(model)

#OLDEST
model <- lmer(pdr_z_filtered ~ region_1s * partner + baseline_pdr + lum_z + (1|subid) + (1|vid), data = filter(bounds_1s, age_group == "oldest"))
summary(model)
anova(model)

model <- lmer(pdr_z_filtered ~ region_1s + baseline_pdr + lum_z + (1|subid) + (1|vid), data = filter(bounds_1s, age_group == "oldest" & partner == "Infant"))
summary(model)
anova(model)

model <- lmer(pdr_z_filtered ~ region_1s + baseline_pdr + lum_z + (1|subid) + (1|vid), data = filter(bounds_1s, age_group == "middlest" & partner == "Adult"))
summary(model)
anova(model)

```
#6. Same analysis controlling for order.
```{r controlling for order}
#read in order file
order <- read_csv(here('chapter_3', 'pdr_orders.csv'))

#create a column that combines mese and video order
order <- order %>% 
  mutate(order = case_when(mese_order == 1 & vid_order == 1 ~ 1,
                           mese_order == 1 & vid_order == 2 ~ 2,
                           mese_order == 2 & vid_order == 1 ~ 3,
                           mese_order == 2 & vid_order == 2 ~ 4))

vid_looking <- left_join(vid_looking, order)

#run analysis examining looking by interaction partner
order_mod <- lmer(prop_looking ~ partner * block + order + (1|subid) + (1|vid), data = vid_looking)
summary(order_mod)
anova(order_mod)

#join with boundary file
bounds_1s <- left_join(bounds_1s, order)

model <- lmer(pdr_z_filtered ~ region_1s * partner + order + baseline_pdr + lum_z + (1|subid) + (1|vid), data = bounds_1s)
summary(model)
anova(model)

#if I add an interaction with order, the interaction is significant, take a look at that
plot.data <- bounds_1s %>% 
  group_by(region_1s, partner, reg_num, order) %>% 
  summarise(mean_pdr = mean(pdr_z_filtered))

rects <- data.frame(xstart = c(-30, 0, 29), xend = c(0, 29, 59), Region = c("pre-boundary", "boundary", "post-boundary"))

rects <- rects %>% 
  mutate(Region = factor(Region, levels = c("pre-boundary", "boundary", "post-boundary")))

ggplot(plot.data) +
  geom_rect(data = rects, aes(xmin = xstart, xmax = xend, ymin = -Inf, ymax = Inf, fill = Region), alpha = 0.2, inherit.aes = FALSE) +
  geom_line(aes(x = reg_num, y = mean_pdr, linetype = partner), size = 2) +
  theme_linedraw(base_size = 16) +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(y = "Pupil Size (Black) & Luminance (Red)", x = "Frame Number (Boundary at 0)", linetype = "Interaction Partner") +
  guides(linetype = guide_legend(override.aes = list(size = 1))) +
  facet_wrap(~order)

```
#6. Same analysis controlling for how many times each object was presented in the pair that it was presented in...
```{r controlling for number of times each pair presented}
#read in pairs file
pairs <- read_csv(here('chapter_3', 'pair_presented.csv'))

vid_looking <- left_join(vid_looking, pairs)

#run analysis examining looking by interaction partner
pairs_mod <- lmer(prop_looking ~ partner * block + pair_presented + (1|subid) + (1|vid), data = vid_looking)
summary(order_mod)
anova(order_mod)

#join with boundary file
bounds_1s <- left_join(bounds_1s, pairs)

model <- lmer(pdr_z_filtered ~ region_1s * partner * pair_presented + baseline_pdr + lum_z + (1|subid) + (1|vid), data = bounds_1s)
summary(model)
anova(model)

#if I add an interaction with order, the interaction is significant, take a look at that
plot.data <- bounds_1s %>% 
  group_by(region_1s, partner, reg_num, order) %>% 
  summarise(mean_pdr = mean(pdr_z_filtered))

rects <- data.frame(xstart = c(-30, 0, 29), xend = c(0, 29, 59), Region = c("pre-boundary", "boundary", "post-boundary"))

rects <- rects %>% 
  mutate(Region = factor(Region, levels = c("pre-boundary", "boundary", "post-boundary")))

ggplot(plot.data) +
  geom_rect(data = rects, aes(xmin = xstart, xmax = xend, ymin = -Inf, ymax = Inf, fill = Region), alpha = 0.2, inherit.aes = FALSE) +
  geom_line(aes(x = reg_num, y = mean_pdr, linetype = partner), size = 2) +
  theme_linedraw(base_size = 16) +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(y = "Pupil Size (Black) & Luminance (Red)", x = "Frame Number (Boundary at 0)", linetype = "Interaction Partner") +
  guides(linetype = guide_legend(override.aes = list(size = 1))) +
  facet_wrap(~order)
```
#EXTRA. Look at PDR for objects when they're paired with other specific objects.
```{r examining by pair}
#GREEN TUBE
GreenTube <- bounds_1s %>% 
  filter(object == "GreenTube")

GreenTube <- GreenTube %>% 
  mutate(pair = paste(pair, "(", pair_presented, ")"))

model <- lmer(pdr_z_filtered ~ region_1s * partner * pair + baseline_pdr + lum_z + (1|subid) + (1|vid), data = GreenTube)
summary(model)
anova(model)

#Plot
plot.data <- GreenTube %>% 
  group_by(region_1s, partner, reg_num, pair) %>% 
  summarise(mean_pdr = mean(pdr_z_filtered))

rects <- data.frame(xstart = c(-30, 0, 29), xend = c(0, 29, 59), Region = c("pre-boundary", "boundary", "post-boundary"))

rects <- rects %>% 
  mutate(Region = factor(Region, levels = c("pre-boundary", "boundary", "post-boundary")))

ggplot(plot.data) +
  geom_rect(data = rects, aes(xmin = xstart, xmax = xend, ymin = -Inf, ymax = Inf, fill = Region), alpha = 0.2, inherit.aes = FALSE) +
  geom_line(aes(x = reg_num, y = mean_pdr, linetype = partner), size = 2) +
  theme_linedraw(base_size = 16) +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(y = "Pupil Size (Black) & Luminance (Red)", x = "Frame Number (Boundary at 0)", linetype = "Interaction Partner") +
  guides(linetype = guide_legend(override.aes = list(size = 1))) +
  facet_wrap(~pair)

#MASSAGEROLLER
MassageRoller <- bounds_1s %>% 
  filter(object == "MassageRoller")

MassageRoller <- MassageRoller %>% 
  mutate(pair = paste(pair, "(", pair_presented, ")"))

model <- lmer(pdr_z_filtered ~ region_1s * partner * pair + baseline_pdr + lum_z + (1|subid) + (1|vid), data = MassageRoller)
summary(model)
anova(model)

#Plot
plot.data <- MassageRoller %>% 
  group_by(region_1s, partner, reg_num, pair) %>% 
  summarise(mean_pdr = mean(pdr_z_filtered))

rects <- data.frame(xstart = c(-30, 0, 29), xend = c(0, 29, 59), Region = c("pre-boundary", "boundary", "post-boundary"))

rects <- rects %>% 
  mutate(Region = factor(Region, levels = c("pre-boundary", "boundary", "post-boundary")))

ggplot(plot.data) +
  geom_rect(data = rects, aes(xmin = xstart, xmax = xend, ymin = -Inf, ymax = Inf, fill = Region), alpha = 0.2, inherit.aes = FALSE) +
  geom_line(aes(x = reg_num, y = mean_pdr, linetype = partner), size = 2) +
  theme_linedraw(base_size = 16) +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(y = "Pupil Size (Black) & Luminance (Red)", x = "Frame Number (Boundary at 0)", linetype = "Interaction Partner") +
  guides(linetype = guide_legend(override.aes = list(size = 1))) +
  facet_wrap(~pair)


```
#EXTRA. Is infants' PDR greater at boundaries than pre- or post-boundary regions?
# 500ms window
```{r PDR to boundary regions .5s}
#import boundary regions file
bounds_half_sec <- read_csv('chapter_3/boundary_regions_500ms_after_b.csv')

#join with PDR file to keep only boundary regions
bounds_half_sec <- left_join(bounds_half_sec, video_pdr)

#add object/partner
bounds_half_sec <- bounds_half_sec %>% 
  separate(vid, into = c("actor", "partner", "object"), sep = "_", remove = FALSE)

bounds_half_sec$region <- factor(bounds_half_sec$region, levels = c("pre_boundary", "boundary", "post_boundary"))

contrasts(bounds_half_sec$region) <- contr.poly

#check that the contrasts look right
attributes(bounds_half_sec$region)

model <- lmer(pdr_z_filtered ~ region + baseline_pdr + (1|subid) + (1|vid), data = bounds_half_sec)
summary(model)
anova(model)

#plot this non-significant effect anyway
plot.data <- bounds_half_sec %>% 
  group_by(reg_num, region) %>% 
  dplyr::summarise(avg_pdr = mean(pdr_z_filtered, na.rm = TRUE))

plot.data %>% 
  ggplot(aes(x = reg_num, y = avg_pdr)) +
  geom_line(size = 1) +
  labs(title = "PDR to Boundaries", x = "Frames; Boundary at 0", y = "Average Z-Scored, Filtered PDR") +
  geom_vline(xintercept = 0, linetype = "dashed", size = 1) +
  geom_vline(xintercept = c(-15, 14, 39), color = "red", size = 1)


```
#EXTRA. Does motionese enhance infant's attention to event boundaries?
# 500ms window
```{r boundary x motionese PDR .5s}
model <- lmer(pdr_z_filtered ~ region * partner + baseline_pdr + (1|subid) + (1|vid), data = bounds_half_sec)
summary(model)
anova(model)

model_adult <- lmer(pdr_z_filtered ~ region + baseline_pdr + (1|subid) + (1|vid), data = filter(bounds_half_sec, partner == "Adult"))
summary(model_adult)
anova(model_adult)

est.marginal.means <- lsmeans(model_adult, specs="region")
est.marginal.means
pairs(est.marginal.means, adjust = "bonferroni")

model_infant <- lmer(pdr_z_filtered ~ region + baseline_pdr + (1|subid) + (1|vid), data = filter(bounds_half_sec, partner == "Infant"))
summary(model_infant)
anova(model_infant)

est.marginal.means <- lsmeans(model_infant, specs="region")
est.marginal.means
pairs(est.marginal.means, adjust = "bonferroni")

bounds_half_sec %>% 
  filter(partner == "Infant") %>% 
  group_by(region) %>% 
  summarise(mean_infant = mean(pdr_z_filtered),
            sd_infant = sd(pdr_z_filtered))

plot.data <- bounds_half_sec %>% 
  group_by(reg_num, region, partner) %>% 
  dplyr::summarise(avg_pdr = mean(pdr_z_filtered, na.rm = TRUE))

plot.data %>% 
  ggplot(aes(x = reg_num, y = avg_pdr, color = partner)) +
  geom_line(size = 1) +
  labs(title = "PDR to Boundaries for Adult- vs Infant-Directed Action", x = "Frames; Boundary at 0", y = "Average Z-Scored, Filtered, Baseline Corrected PDR") +
  geom_vline(xintercept = 0, size = 1) +
  geom_vline(xintercept = c(-15, 14, 29), linetype = "dashed", size = 1)

#plot the same way as I did luminance
plot.data <- bounds_half_sec %>% 
  group_by(region, partner, reg_num) %>% 
  summarise(mean = mean(pdr_z_filtered))

rects <- data.frame(xstart = c(-15, 0, 14), xend = c(0, 14, 29), Region = c("pre-boundary", "boundary", "post-boundary"))

rects <- rects %>% 
  mutate(Region = factor(Region, levels = c("pre-boundary", "boundary", "post-boundary")))

ggplot(plot.data, aes(x = reg_num, y = mean, linetype = partner)) +
  geom_rect(data = rects, aes(xmin = xstart, xmax = xend, ymin = -Inf, ymax = Inf, fill = Region), alpha = 0.2, inherit.aes = FALSE) +
  geom_line(size = 2) +
  theme_linedraw(base_size = 16) +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(y = "Average Z-Scored, Filtered PDR", x = "Frame Number (Boundary at 0)", linetype = "Interaction Partner")

```
#corrected version of luminance
```{r}
luminance_corr <- read_csv('corrected_luminance_7.15.19.csv')

#remove subid from vid in luminance_corr file
luminance_corr <- luminance_corr %>% 
  separate(vid, c("partner", "object")) 

#select only the columns we want to join with the data file
luminance_corr <- luminance_corr[,c("partner", "object", "frame_num", "lum_z")]

#join with boundary file
bounds_1s <- left_join(bounds_1s, luminance_corr)

model <- lmer(pdr_z_filtered ~ region_1s * partner + baseline_pdr + lum_z + (1|subid) + (1|vid), data = bounds_1s)
summary(model)
anova(model)

model_adult <- lmer(pdr_z_filtered ~ region_1s + baseline_pdr + lum_z + (1|subid) + (1|vid), data = filter(bounds_1s, partner == "Adult"))
summary(model_adult)
anova(model_adult)

model_infant <- lmer(pdr_z_filtered ~ region_1s + baseline_pdr + lum_z + (1|subid) + (1|vid), data = filter(bounds_1s, partner == "Infant"))
summary(model_infant)
anova(model_infant)

#plot with luminance_corr on same plot as PDR
#plot the same way as I did luminance_corr
plot.data <- bounds_1s %>% 
  group_by(region_1s, partner, reg_num) %>% 
  summarise(mean_pdr = mean(pdr_z_filtered),
            mean_lum = mean(lum_z))

rects <- data.frame(xstart = c(-30, 0, 29), xend = c(0, 29, 59), Region = c("pre-boundary", "boundary", "post-boundary"))

rects <- rects %>% 
  mutate(Region = factor(Region, levels = c("pre-boundary", "boundary", "post-boundary")))

ggplot(plot.data) +
  geom_rect(data = rects, aes(xmin = xstart, xmax = xend, ymin = -Inf, ymax = Inf, fill = Region), alpha = 0.2, inherit.aes = FALSE) +
  geom_line(aes(x = reg_num, y = mean_pdr, linetype = partner), size = 2) +
  geom_line(aes(x = reg_num, y = mean_lum, linetype = partner), color = "red", size = 2) +
  theme_linedraw(base_size = 16) +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(y = "Average Z-Scored, Filtered PDR", x = "Frame Number (Boundary at 0)", linetype = "Interaction Partner")

```
# Write data file for manuscript.
```{r write data file}
write_csv(bounds_1s, "cleaned_pdr_data.csv")
write_csv(vid_looking, "cleaned_looking_data.csv")
```
